Change to read unit from preferences

+++ accudrop/app/src/main/java/me/chrislane/accudrop/Point3D.java
@@ -16,7 +16,7 @@ public class Point3D {
         return latLng;
     }
 
-    public double getAltitude() {
+    public float getAltitude() {
         return altitude;
     }
